exploding the EventDetails's components in multiple files for clarity
